requirements.txt

Flask==1.1.4
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.23
Flask-Migrate==2.5.3
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.25.1
python-dotenv==0.19.2
markupsafe==2.0.1



